refactor: pull `CanonicalPath` out of the environment module
