+++ cli/flox-rust-sdk/src/data/mod.rs
@@ -1,4 +1,6 @@
+mod canonical_path;
 mod version;
 
+pub use canonical_path::{CanonicalPath, CanonicalizeError};
 pub use version::Version;
 pub type System = String;
